fix(view): getDirection should return null if segments do not match
BREAKING CHANGE:
getDirection method is now more strict about detecting the direction from URL segments.
It requires the URLs to be the same in but the one segment that they defer from one
another. Examples below illustrate URL changes and the returned value of getDirection
if fed with the respective URL segments.

Before:
/home -> /home/second returned 'normal'
/home/second -> /home returned 'reverse'
/home/second/third -> /home returned null
/home -> /home/second/third returned 'normal'
/home/second -> /crazy/bar/baz returned 'normal'

After:
/home -> /home/second returns 'normal'
/home/second -> /home returns 'reverse'
/home/second/third -> /home returns null
/home -> /home/second/third returns null
/home/second -> /crazy/bar/baz returns null