fix(bpTap): remove bpTap, replace w/ ngClick
bpTap is a very specialiced directive that closely resembles native touch behavior (esp. on iOS devices).
Nonetheless it is not consistent to duplicate core functionality of AngularJS, rather than overwriting it.
Simply replacing bpTap now is just a first step towards compatibility and we'll bring back the fine-grained control of bpTap,
either by comitting directly to AngularJS, or overwriting the directive in the future.

BREAKING CHANGE: bpTap is now named ngClick

before:
```html
<bp-action bp-tap="doSomething()"></bp-action>
```

after:
```html
<bp-action ng-click="doSomething()"></bp-action>
```

This should be a search and replace migration.